backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "enerbrain-backend"
version = "0.1.0"
description = "EnerBrain FastAPI 백엔드"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "PyMySQL",
    "psycopg[binary]>=3.2.0",
    "PyJWT>=2.9.0",
    "passlib[argon2]>=1.7.4",
]

[project.scripts]
enerbrain-serve = "app.cli:main"

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
